Move MatchaStateService to matcha_ml.state
