+++ src/matcha_ml/core/core.py
@@ -4,7 +4,7 @@ from typing import Dict, Optional
 
 from matcha_ml.cli._validation import get_command_validation
 from matcha_ml.errors import MatchaError
-from matcha_ml.services.matcha_state import MatchaStateService
+from matcha_ml.state import MatchaStateService
 
 
 def get(
