+++ src/matcha_ml/state/__init__.py
@@ -1,4 +1,5 @@
 """Matcha state sub-package."""
+from .matcha_state import MatchaStateService
 from .remote_state_manager import RemoteStateManager
 
-__all__ = ["RemoteStateManager"]
+__all__ = ["RemoteStateManager", "MatchaStateService"]
