+++ src/matcha_ml/templates/build_templates/azure_template.py
@@ -16,7 +16,7 @@ from matcha_ml.cli.ui.status_message_builders import (
     build_substep_success_status,
 )
 from matcha_ml.errors import MatchaPermissionError
-from matcha_ml.services.matcha_state import MatchaStateService
+from matcha_ml.state import MatchaStateService
 
 SUBMODULE_NAMES = [
     "aks",
