+++ tests/test_services/test_matcha_state.py
@@ -4,7 +4,7 @@ import os
 
 import pytest
 
-from matcha_ml.services.matcha_state import MatchaStateService
+from matcha_ml.state import MatchaStateService
 
 
 @pytest.fixture
